+++ R/prop_diff.R
@@ -8,6 +8,21 @@
 #' constructing the confidence interval of the proportion difference. A stratification variable can be supplied via the
 #' `strata` element of the `variables` argument.
 #'
+#' @details The possible methods are:
+#'
+#' - `"waldcc"`: Wald confidence interval with continuity correction \insertCite{Agresti1998}{tern}.
+#' - `"wald"`: Wald confidence interval without continuity correction \insertCite{Agresti1998}{tern}.
+#' - `"cmh"`: Cochran-Mantel-Haenszel (CMH) confidence interval \insertCite{MantelHaenszel1959}{tern}.
+#' - `"cmh_sato"`: CMH confidence interval with Sato variance estimator \insertCite{Sato1989}{tern}.
+#' - `"cmh_mn"`: CMH confidence interval with Miettinen and Nurminen confidence interval
+#'      \insertCite{MiettinenNurminen1985}{tern}.
+#' - `"ha"`: Anderson-Hauck confidence interval \insertCite{HauckAnderson1986}{tern}.
+#' - `"newcombe"`: Newcombe confidence interval without continuity correction \insertCite{Newcombe1998}{tern}.
+#' - `"newcombecc"`: Newcombe confidence interval with continuity correction \insertCite{Newcombe1998}{tern}.
+#' - `"strat_newcombe"`: Stratified Newcombe confidence interval without continuity
+#'     correction \insertCite{Yan2010-jt}{tern}.
+#' - `"strat_newcombecc"`: Stratified Newcombe confidence interval with continuity
+#'     correction \insertCite{Yan2010-jt}{tern}.
 #'
 #' @inheritParams prop_diff_strat_nc
 #' @inheritParams argument_convention
@@ -17,6 +32,9 @@
 #'   Options are: ``r shQuote(get_stats("estimate_proportion_diff"), type = "sh")``
 #'
 #' @seealso [d_proportion_diff()]
+#'
+#' @references
+#'   \insertAllCited{}
 #'
 #' @name prop_diff
 #' @order 1
@@ -28,8 +46,8 @@ NULL
 #' @return
 #' * `s_proportion_diff()` returns a named list of elements `diff` and `diff_ci`.
 #'
-#' @note When performing an unstratified analysis, methods `"cmh"`, `"strat_newcombe"`, and `"strat_newcombecc"` are
-#'   not permitted.
+#' @note When performing an unstratified analysis, methods `"cmh"`, `"cmh_sato"`, `"strat_newcombe"`,
+#'   and `"strat_newcombecc"` are not permitted.
 #'
 #' @examples
 #' s_proportion_diff(
@@ -60,16 +78,20 @@ s_proportion_diff <- function(df,
                               variables = list(strata = NULL),
                               conf_level = 0.95,
                               method = c(
-                                "waldcc", "wald", "cmh",
+                                "waldcc", "wald", "cmh", "cmh_sato", "cmh_mn",
                                 "ha", "newcombe", "newcombecc",
                                 "strat_newcombe", "strat_newcombecc"
                               ),
                               weights_method = "cmh",
                               ...) {
   method <- match.arg(method)
-  if (is.null(variables$strata) && checkmate::test_subset(method, c("cmh", "strat_newcombe", "strat_newcombecc"))) {
+  if (
+    is.null(variables$strata) &&
+      checkmate::test_subset(method, c("cmh", "cmh_sato", "cmh_mn", "strat_newcombe", "strat_newcombecc"))
+  ) {
     stop(paste(
-      "When performing an unstratified analysis, methods 'cmh', 'strat_newcombe', and 'strat_newcombecc' are not",
+      "When performing an unstratified analysis, methods",
+      "'cmh', 'cmh_sato', 'cmh_mn', 'strat_newcombe', and 'strat_newcombecc' are not",
       "permitted. Please choose a different method."
     ))
   }
@@ -128,7 +150,9 @@ s_proportion_diff <- function(df,
         conf_level,
         correct = TRUE
       ),
-      "cmh" = prop_diff_cmh(rsp, grp, strata, conf_level)[c("diff", "diff_ci")]
+      "cmh" = prop_diff_cmh(rsp, grp, strata, conf_level, diff_se = "standard")[c("diff", "diff_ci")],
+      "cmh_sato" = prop_diff_cmh(rsp, grp, strata, conf_level, diff_se = "sato")[c("diff", "diff_ci")],
+      "cmh_mn" = prop_diff_cmh(rsp, grp, strata, conf_level, diff_se = "miettinen_nurminen")[c("diff", "diff_ci")]
     )
 
     y$diff <- setNames(y$diff * 100, paste0("diff_", method))
@@ -266,7 +290,7 @@ estimate_proportion_diff <- function(lyt,
                                      variables = list(strata = NULL),
                                      conf_level = 0.95,
                                      method = c(
-                                       "waldcc", "wald", "cmh",
+                                       "waldcc", "wald", "cmh", "cmh_sato", "cmh_mn",
                                        "ha", "newcombe", "newcombecc",
                                        "strat_newcombe", "strat_newcombecc"
                                      ),
@@ -387,7 +411,7 @@ d_proportion_diff <- function(conf_level,
     label <- paste(
       label,
       ifelse(
-        method == "cmh",
+        method %in% c("cmh", "cmh_sato", "cmh_mn"),
         "for adjusted difference",
         "for difference"
       )
@@ -396,6 +420,8 @@ d_proportion_diff <- function(conf_level,
 
   method_part <- switch(method,
     "cmh" = "CMH, without correction",
+    "cmh_sato" = "CMH, Sato variance estimator",
+    "cmh_mn" = "CMH, Miettinen and Nurminen",
     "waldcc" = "Wald, with correction",
     "wald" = "Wald, without correction",
     "ha" = "Anderson-Hauck",
@@ -421,6 +447,9 @@ d_proportion_diff <- function(conf_level,
 #'   (proportion difference confidence interval).
 #'
 #' @seealso [prop_diff()] for implementation of these helper functions.
+#'
+#' @references
+#'   \insertAllCited{}
 #'
 #' @name h_prop_diff
 NULL
@@ -475,7 +504,7 @@ prop_diff_wald <- function(rsp,
   )
 }
 
-#' @describeIn h_prop_diff Anderson-Hauck confidence interval.
+#' @describeIn h_prop_diff Anderson-Hauck confidence interval \insertCite{HauckAnderson1986}{tern}.
 #'
 #' @examples
 #' # Anderson-Hauck confidence interval
@@ -513,7 +542,7 @@ prop_diff_ha <- function(rsp,
 }
 
 #' @describeIn h_prop_diff Newcombe confidence interval. It is based on
-#'   the Wilson score confidence interval for a single binomial proportion.
+#'   the Wilson score confidence interval for a single binomial proportion \insertCite{Newcombe1998}{tern}.
 #'
 #' @examples
 #' # Newcombe confidence interval
@@ -563,6 +592,9 @@ prop_diff_nc <- function(rsp,
 #'   test, use [stats::mantelhaen.test()].
 #'
 #' @param strata (`factor`)\cr variable with one level per stratum and same length as `rsp`.
+#' @param diff_se (`string`)\cr method to estimate the standard error for the difference, either
+#'   `standard`, `sato` \insertCite{Sato1989}{tern} or
+#'   `miettinen_nurminen` \insertCite{MiettinenNurminen1985}{tern}.
 #'
 #' @examples
 #' # Cochran-Mantel-Haenszel confidence interval
@@ -581,14 +613,20 @@ prop_diff_nc <- function(rsp,
 #'   rsp = rsp, grp = grp, strata = interaction(strata_data),
 #'   conf_level = 0.90
 #' )
+#' prop_diff_cmh(
+#'   rsp = rsp, grp = grp, strata = interaction(strata_data),
+#'   conf_level = 0.90, diff_se = "sato"
+#' )
 #'
 #' @export
 prop_diff_cmh <- function(rsp,
                           grp,
                           strata,
-                          conf_level = 0.95) {
+                          conf_level = 0.95,
+                          diff_se = c("standard", "sato", "miettinen_nurminen")) {
   grp <- as_factor_keep_attributes(grp)
   strata <- as_factor_keep_attributes(strata)
+  diff_se <- match.arg(diff_se)
   check_diff_prop_ci(
     rsp = rsp, grp = grp, conf_level = conf_level, strata = strata
   )
@@ -608,8 +646,10 @@ prop_diff_cmh <- function(rsp,
   )
   n1 <- colSums(t_tbl[1:2, 1, ])
   n2 <- colSums(t_tbl[1:2, 2, ])
-  p1 <- t_tbl[2, 1, ] / n1
-  p2 <- t_tbl[2, 2, ] / n2
+  x1 <- t_tbl[2, 1, ]
+  p1 <- x1 / n1
+  x2 <- t_tbl[2, 2, ]
+  p2 <- x2 / n2
   # CMH weights
   use_stratum <- (n1 > 0) & (n2 > 0)
   n1 <- n1[use_stratum]
@@ -632,17 +672,117 @@ prop_diff_cmh <- function(rsp,
   estimate_ci <- list(ci1, ci2)
   names(estimate_ci) <- levels(grp)
   diff_est <- est2 - est1
-  se_diff <- sqrt(sum(((p1 * (1 - p1) / n1) + (p2 * (1 - p2) / n2)) * wt_normalized^2))
-  diff_ci <- c(diff_est - z * se_diff, diff_est + z * se_diff)
+
+  if (diff_se %in% c("standard", "sato")) {
+    se_diff <- if (diff_se == "standard") {
+      sqrt(sum(((p1 * (1 - p1) / n1) + (p2 * (1 - p2) / n2)) * wt_normalized^2))
+    } else {
+      # Sato variance estimator.
+      p_terms <- (n2^2 * x1 - n1^2 * x2 + n1 * n2 * (n1 - n2) / 2) / (n1 + n2)^2
+      q_terms <- (x1 * (n2 - x2) + x2 * (n1 - x1)) / (2 * (n1 + n2))
+      num <- diff_est * sum(p_terms) + sum(q_terms)
+      denom <- sum(wt)^2
+      sqrt(num / denom)
+    }
+    diff_ci <- c(diff_est - z * se_diff, diff_est + z * se_diff)
+  } else {
+    # Miettinen and Nurminen method is used.
+    z_stat_fun <- function(delta) {
+      var_est <- h_miettinen_nurminen_var_est(
+        n1 = n1, n2 = n2,
+        x1 = x1, x2 = x2,
+        diff_par = delta
+      )$var_est
+      num <- sum(wt * (p2 - p1 - delta))
+      denom <- sqrt(sum(wt^2 * var_est))
+      num / denom
+    }
+    # Find upper and lower confidence limits by root finding such that
+    # z_stat_fun(limit) = +/- z quantile:
+    root_lower <- function(delta) z_stat_fun(delta) - z
+    root_upper <- function(delta) z_stat_fun(delta) + z
+    diff_ci <- c(
+      stats::uniroot(root_lower, interval = c(-0.99, diff_est))$root,
+      stats::uniroot(root_upper, interval = c(diff_est, 0.99))$root
+    )
+    # Calculate the standard error separately.
+    var_est <- h_miettinen_nurminen_var_est(
+      n1 = n1, n2 = n2,
+      x1 = x1, x2 = x2,
+      diff_par = diff_est
+    )$var_est
+    se_diff <- sqrt(sum(wt_normalized^2 * var_est))
+  }
 
   list(
     prop = estimate,
     prop_ci = estimate_ci,
     diff = diff_est,
     diff_ci = diff_ci,
+    se_diff = se_diff,
     weights = wt_normalized,
     n1 = n1,
     n2 = n2
+  )
+}
+
+#' Variance Estimates in Strata following Miettinen and Nurminen
+#'
+#' The variable names in this function follow the notation in the original
+#' paper by \insertCite{MiettinenNurminen1985;textual}{tern}, cf. Appendix 1.
+#'
+#' @param n1 (`numeric`)\cr sample sizes in group 1.
+#' @param n2 (`numeric`)\cr sample sizes in group 2.
+#' @param x1 (`numeric`)\cr number of responders in group 1.
+#' @param x2 (`numeric`)\cr number of responders in group 2.
+#' @param diff_par (`numeric`)\cr assumed difference in true proportions
+#'   (group 2 minus group 1).
+#' @return A named `list` with elements:
+#'
+#' - `p1_hat`: estimated proportion in group 1
+#' - `p2_hat`: estimated proportion in group 2
+#' - `var_est`: variance estimate of the difference in proportions
+#'
+#' @keywords internal
+#' @references
+#'   \insertAllCited{}
+h_miettinen_nurminen_var_est <- function(n1, n2, x1, x2, diff_par) {
+  # nolint start
+  # Translate to the notation in the paper.
+  S0 <- n1
+  S1 <- n2
+  c0 <- x1
+  c1 <- x2
+  RD <- diff_par
+
+  # Further definitions.
+  S <- S0 + S1
+  c <- c0 + c1
+
+  # Coefficients of the third-degree polynomial.
+  L3 <- S
+  L2 <- (S1 + 2 * S0) * RD - S - c
+  L1 <- (S0 * RD - S - 2 * c0) * RD + c
+  L0 <- c0 * RD * (1 - RD)
+  # nolint end
+
+  # Solution for group 1 proportion.
+  q <- L2^3 / (3 * L3)^3 - L1 * L2 / (6 * L3^2) + L0 / (2 * L3)
+  p <- sign(q) * sqrt(L2^2 / (3 * L3)^2 - L1 / (3 * L3))
+  a <- (1 / 3) * (base::pi + acos(q / p^3))
+  p1_hat <- 2 * p * cos(a) - L2 / (3 * L3)
+
+  # Estimated group 2 proportion.
+  p2_hat <- p1_hat + RD
+
+  # Variance estimate.
+  var_est <- (p1_hat * (1 - p1_hat) / n1 + p2_hat * (1 - p2_hat) / n2) *
+    S / (S - 1)
+
+  list(
+    p1_hat = p1_hat,
+    p2_hat = p2_hat,
+    var_est = var_est
   )
 }
 
@@ -655,9 +795,6 @@ prop_diff_cmh <- function(rsp,
 #' @param strata (`factor`)\cr variable with one level per stratum and same length as `rsp`.
 #' @param weights_method (`string`)\cr weights method. Can be either `"cmh"` or `"heuristic"`
 #'   and directs the way weights are estimated.
-#'
-#' @references
-#' \insertRef{Yan2010-jt}{tern}
 #'
 #' @examples
 #' # Stratified Newcombe confidence interval