+++ R/prop_diff_test.R
@@ -4,12 +4,13 @@
 #'
 #' The analyze function [test_proportion_diff()] creates a layout element to test the difference between two
 #' proportions. The primary analysis variable, `vars`, indicates whether a response has occurred for each record. See
-#' the `method` parameter for options of methods to use to calculate the p-value. Additionally, a stratification
-#' variable can be supplied via the `strata` element of the `variables` argument.
+#' the `method` parameter for options of methods to use to calculate the p-value. The argument `alternative`
+#' specifies the direction of the alternative hypothesis. Additionally, a stratification variable can be
+#' supplied via the `strata` element of the `variables` argument.
 #'
 #' @inheritParams argument_convention
-#' @param method (`string`)\cr one of `chisq`, `cmh`, `fisher`, or `schouten`; specifies the test used
-#'   to calculate the p-value.
+#' @param method (`string`)\cr one of `chisq`, `cmh`, `cmh_wh`, `fisher`, or `schouten`;
+#'   specifies the test used to calculate the p-value.
 #' @param .stats (`character`)\cr statistics to select for the table.
 #'
 #'   Options are: ``r shQuote(get_stats("test_proportion_diff"), type = "sh")``
@@ -52,7 +53,8 @@ s_test_proportion_diff <- function(df,
                                    .ref_group,
                                    .in_ref_col,
                                    variables = list(strata = NULL),
-                                   method = c("chisq", "schouten", "fisher", "cmh"),
+                                   method = c("chisq", "schouten", "fisher", "cmh", "cmh_wh"),
+                                   alternative = c("two.sided", "less", "greater"),
                                    ...) {
   method <- match.arg(method)
   y <- list(pval = numeric())
@@ -69,7 +71,7 @@ s_test_proportion_diff <- function(df,
       levels = c("ref", "Not-ref")
     )
 
-    if (!is.null(variables$strata) || method == "cmh") {
+    if (!is.null(variables$strata) || method %in% c("cmh", "cmh_wh")) {
       strata <- variables$strata
       checkmate::assert_false(is.null(strata))
       strata_vars <- stats::setNames(as.list(strata), strata)
@@ -80,18 +82,20 @@ s_test_proportion_diff <- function(df,
 
     tbl <- switch(method,
       cmh = table(grp, rsp, strata),
+      cmh_wh = table(grp, rsp, strata),
       table(grp, rsp)
     )
 
     y$pval <- switch(method,
-      chisq = prop_chisq(tbl),
-      cmh = prop_cmh(tbl),
-      fisher = prop_fisher(tbl),
-      schouten = prop_schouten(tbl)
+      chisq = prop_chisq(tbl, alternative = alternative),
+      cmh = prop_cmh(tbl, alternative = alternative),
+      fisher = prop_fisher(tbl, alternative = alternative),
+      schouten = prop_schouten(tbl, alternative = alternative),
+      cmh_wh = prop_cmh(tbl, alternative = alternative, transform = "wilson_hilferty")
     )
   }
 
-  y$pval <- formatters::with_label(y$pval, d_test_proportion_diff(method))
+  y$pval <- formatters::with_label(y$pval, d_test_proportion_diff(method, alternative = alternative))
   y
 }
 
@@ -106,16 +110,24 @@ s_test_proportion_diff <- function(df,
 #' @return A `string` describing the test from which the p-value is derived.
 #'
 #' @export
-d_test_proportion_diff <- function(method) {
+d_test_proportion_diff <- function(method, alternative = c("two.sided", "less", "greater")) {
   checkmate::assert_string(method)
+  alternative <- match.arg(alternative)
+
   meth_part <- switch(method,
     "schouten" = "Chi-Squared Test with Schouten Correction",
     "chisq" = "Chi-Squared Test",
     "cmh" = "Cochran-Mantel-Haenszel Test",
+    "cmh_wh" = "Cochran-Mantel-Haenszel Test with Wilson-Hilferty Transformation",
     "fisher" = "Fisher's Exact Test",
     stop(paste(method, "does not have a description"))
   )
-  paste0("p-value (", meth_part, ")")
+  alt_part <- switch(alternative,
+    two.sided = "",
+    less = ", 1-sided, direction less",
+    greater = ", 1-sided, direction greater"
+  )
+  paste0("p-value (", meth_part, alt_part, ")")
 }
 
 #' @describeIn prop_diff_test Formatted analysis function which is used as `afun` in `test_proportion_diff()`.
@@ -223,7 +235,8 @@ a_test_proportion_diff <- function(df,
 test_proportion_diff <- function(lyt,
                                  vars,
                                  variables = list(strata = NULL),
-                                 method = c("chisq", "schouten", "fisher", "cmh"),
+                                 method = c("chisq", "schouten", "fisher", "cmh", "cmh_wh"),
+                                 alternative = c("two.sided", "less", "greater"),
                                  var_labels = vars,
                                  na_str = default_na_str(),
                                  nested = TRUE,
@@ -242,6 +255,7 @@ test_proportion_diff <- function(lyt,
     "na_rm" = na_rm,
     "variables" = variables,
     "method" = method,
+    "alternative" = alternative,
     ...
   )
 
@@ -280,6 +294,7 @@ test_proportion_diff <- function(lyt,
 #' Helper functions to implement various tests on the difference between two proportions.
 #'
 #' @param tbl (`matrix`)\cr matrix with two groups in rows and the binary response (`TRUE`/`FALSE`) in columns.
+#' @inheritParams argument_convention
 #'
 #' @return A p-value.
 #'
@@ -291,33 +306,57 @@ NULL
 #' @describeIn h_prop_diff_test Performs Chi-Squared test. Internally calls [stats::prop.test()].
 #'
 #' @keywords internal
-prop_chisq <- function(tbl) {
+prop_chisq <- function(tbl, alternative = c("two.sided", "less", "greater")) {
   checkmate::assert_integer(c(ncol(tbl), nrow(tbl)), lower = 2, upper = 2)
   tbl <- tbl[, c("TRUE", "FALSE")]
   if (any(colSums(tbl) == 0)) {
     return(1)
   }
-  stats::prop.test(tbl, correct = FALSE)$p.value
+  stats::prop.test(tbl, correct = FALSE, alternative = alternative)$p.value
 }
 
-#' @describeIn h_prop_diff_test Performs stratified Cochran-Mantel-Haenszel test. Internally calls
-#'   [stats::mantelhaen.test()]. Note that strata with less than two observations are automatically discarded.
+#' @describeIn h_prop_diff_test Performs stratified Cochran-Mantel-Haenszel test,
+#'   using [stats::mantelhaen.test()] internally.
+#'   Note that strata with less than two observations are automatically discarded.
 #'
 #' @param ary (`array`, 3 dimensions)\cr array with two groups in rows, the binary response
 #'   (`TRUE`/`FALSE`) in columns, and the strata in the third dimension.
+#' @param transform (`string`)\cr either `none` or `wilson_hilferty`; specifies whether to apply
+#'   the Wilson-Hilferty transformation of the chi-squared statistic.
 #'
 #' @keywords internal
-prop_cmh <- function(ary) {
+prop_cmh <- function(ary,
+                     alternative = c("two.sided", "less", "greater"),
+                     transform = c("none", "wilson_hilferty")) {
   checkmate::assert_array(ary)
   checkmate::assert_integer(c(ncol(ary), nrow(ary)), lower = 2, upper = 2)
   checkmate::assert_integer(length(dim(ary)), lower = 3, upper = 3)
+  alternative <- match.arg(alternative)
+  transform <- match.arg(transform)
+
   strata_sizes <- apply(ary, MARGIN = 3, sum)
   if (any(strata_sizes < 5)) {
     warning("<5 data points in some strata. CMH test may be incorrect.")
     ary <- ary[, , strata_sizes > 1]
   }
 
-  stats::mantelhaen.test(ary, correct = FALSE)$p.value
+  cmh_res <- stats::mantelhaen.test(ary, correct = FALSE, alternative = alternative)
+
+  if (transform == "none") {
+    cmh_res$p.value
+  } else {
+    chisq_stat <- unname(cmh_res$statistic)
+    df <- unname(cmh_res$parameter)
+    num <- (chisq_stat / df)^(1 / 3) - (1 - 2 / (9 * df))
+    denom <- sqrt(2 / (9 * df))
+    wh_stat <- num / denom
+
+    if (alternative == "two.sided") {
+      2 * stats::pnorm(-abs(wh_stat))
+    } else {
+      stats::pnorm(wh_stat, lower.tail = (alternative == "greater"))
+    }
+  }
 }
 
 #' @describeIn h_prop_diff_test Performs the Chi-Squared test with Schouten correction.
@@ -325,8 +364,9 @@ prop_cmh <- function(ary) {
 #' @seealso Schouten correction is based upon \insertCite{Schouten1980-kd;textual}{tern}.
 #'
 #' @keywords internal
-prop_schouten <- function(tbl) {
+prop_schouten <- function(tbl, alternative = c("two.sided", "less", "greater")) {
   checkmate::assert_integer(c(ncol(tbl), nrow(tbl)), lower = 2, upper = 2)
+  alternative <- match.arg(alternative)
   tbl <- tbl[, c("TRUE", "FALSE")]
   if (any(colSums(tbl) == 0)) {
     return(1)
@@ -345,14 +385,24 @@ prop_schouten <- function(tbl) {
     (abs(prod(ad) - prod(bc)) - 0.5 * min(n1, n2))^2 /
     (n1 * n2 * sum(ac) * sum(bd))
 
-  1 - stats::pchisq(t_schouten, df = 1)
+  if (alternative == "two.sided") {
+    stats::pchisq(t_schouten, df = 1, lower.tail = FALSE)
+  } else {
+    # This follows the logic in stats::prop.test for one-sided p-values.
+    x1 <- tbl[1, 1]
+    x2 <- tbl[2, 1]
+    delta <- (x1 / n1) - (x2 / n2)
+    z <- sign(delta) * sqrt(t_schouten)
+    stats::pnorm(z, lower.tail = (alternative == "less"))
+  }
 }
 
 #' @describeIn h_prop_diff_test Performs the Fisher's exact test. Internally calls [stats::fisher.test()].
 #'
 #' @keywords internal
-prop_fisher <- function(tbl) {
+prop_fisher <- function(tbl, alternative = c("two.sided", "less", "greater")) {
   checkmate::assert_integer(c(ncol(tbl), nrow(tbl)), lower = 2, upper = 2)
+  alternative <- match.arg(alternative) # Is needed here, because stats::fisher.test does not handle defaults.
   tbl <- tbl[, c("TRUE", "FALSE")]
-  stats::fisher.test(tbl)$p.value
+  stats::fisher.test(tbl, alternative = alternative)$p.value
 }