+++ tests/testthat/test-analyze_variables.R
@@ -64,9 +64,9 @@ testthat::test_that("s_summary works when factors have NA levels", {
 
 testthat::test_that("s_summary works with factors with NA values handled and correctly removes them by default", {
   x <- factor(c("Female", "Male", "Female", "Male", "Male", "Unknown", "Unknown", "Unknown", "Unknown", NA))
-  x <- explicit_na(x)
+  x <- explicit_na(x, label = "<Missing>")
 
-  result <- s_summary(x)
+  result <- s_summary(x, na_str_drop = "<Missing>")
 
   res <- testthat::expect_silent(result)
   testthat::expect_snapshot(res)
@@ -682,4 +682,133 @@ testthat::test_that("analyze_vars warnings for geom_verbose work", {
 
   # Do we expect output to be NA?
   expect_true(all(is.na(cell_values(result2)[[1]])))
+})
+
+testthat::test_that("analyze_vars works with all-NA factor input", {
+  data_mtcars <- mtcars
+  data_mtcars$hp2 <- as.factor("<Missing>")
+
+  res <- basic_table() %>%
+    analyze_vars(
+      vars = c("hp2"),
+      .stats = c("n", "count_fraction", "count"),
+      na_rm = TRUE
+    ) %>%
+    build_table(data_mtcars)
+
+  expect_identical(
+    strsplit(toString(matrix_form(res), hsep = "-"), "\n")[[1]],
+    c(
+      "    all obs",
+      "-----------",
+      "n      0   "
+    )
+  )
+})
+
+testthat::test_that("analyze_vars works with all-NA character input", {
+  data_mtcars <- mtcars
+  data_mtcars$hp2 <- "<Missing>"
+
+  res <- basic_table() %>%
+    analyze_vars(
+      vars = c("hp2"),
+      .stats = c("n", "count_fraction", "count"),
+      na_rm = TRUE,
+      verbose = FALSE
+    ) %>%
+    build_table(data_mtcars)
+
+  expect_identical(
+    strsplit(toString(matrix_form(res), hsep = "-"), "\n")[[1]],
+    c(
+      "    all obs",
+      "-----------",
+      "n      0   "
+    )
+  )
+})
+
+test_that("`analyze_vars` works with new rtables format/na_str options in analyze()", {
+  v1fmts <- list(
+    n = "N=xx",
+    mean_sd = "xx.x (xx.x)",
+    median = "xx.x",
+    range = "xx.x - xx.x"
+  )
+
+  v2fmts <- list(
+    n = "(N=xx)",
+    mean_sd = "xx.xx (xx.xx)",
+    median = "xx.xx",
+    range = "xx.xx - xx.xx"
+  )
+
+  v3fmts <- list(
+    n = "xx",
+    mean_sd = "xx. (xx.)",
+    median = "xx.",
+    range = "xx - xx"
+  )
+  dta <- data.frame(
+    USUBJID = rep(1:6, each = 3),
+    AVISIT  = rep(paste0("V", 1:3), 6),
+    ARM     = rep(LETTERS[1:3], rep(6, 3)),
+    AVAL    = c(9:1, rep(NA, 9)),
+    formats = I(rep(list(v1fmts, v2fmts, v3fmts), 6)),
+    na_strs = rep(c("what?", "OK", "-"), 6)
+  )
+
+  lyt <- basic_table() %>%
+    split_cols_by(var = "ARM") %>%
+    split_rows_by(var = "AVISIT") %>%
+    analyze_vars(
+      vars = "AVAL",
+      .formats = "default",
+      formats_var = "formats"
+    )
+
+  tbl <- build_table(lyt, dta)
+
+  fmts <- mf_formats(matrix_form(tbl))[, -1] ## Just Say No To Row Labels (tm)
+
+  fmts_col_vec <- unlist(c("", "-", v1fmts, "-", v2fmts, "-", v3fmts))
+
+  expect_identical(
+    fmts,
+    matrix(fmts_col_vec, nrow = 16, ncol = 3)
+  )
+
+
+  lyt2 <- basic_table() %>%
+    split_cols_by(var = "ARM") %>%
+    split_rows_by(var = "AVISIT") %>%
+    analyze_vars(
+      vars = "AVAL",
+      .formats = "default",
+      formats_var = "formats",
+      na_str = NA,
+      na_strs_var = "na_strs"
+    )
+
+  tbl2 <- build_table(lyt2, dta)
+  fmtcells2 <- get_formatted_cells(tbl2)
+  expect_identical(
+    fmtcells2[, 3, drop = TRUE],
+    c(
+      "",
+      "N=0",
+      rep("what?", 3),
+      "",
+      "(N=0)",
+      rep("OK", 3),
+      "",
+      "0",
+      rep("-", 3)
+    )
+  )
+
+  expect_error(analyze_vars(basic_table(), "AVAL", formats_var = "formats"))
+  expect_error(analyze_vars(basic_table(), "AVAL", na_strs_var = "formats"))
+  expect_error(analyze_vars(basic_table(), "AVAL", format = list(stuff = "in here")))
 })